+++ R/data.R
@@ -41,6 +41,28 @@
 #' }
 "human_suspected"
 
+#' Known and suspected 2'Ome positions in Humans' rRNA
+#'
+#' A dataset containing 129 2'Ome positions with associated data
+#' 
+#' @usage data(human_both_methylated_and_suspected)
+#' 
+#' @format A data frame with 129 rows and 10 variables:
+#' \describe{
+#'   \item{Position}{nucleotide position on rRNA}
+#'   \item{rRNA}{rRNA where this methylation is found}
+#'   \item{Nomenclature}{name given to this modification}
+#'   \item{NR_046235.Numbering}{Position on}
+#'   \item{Sequence}{nucleotides sequence surround position}
+#'   \item{SNORD}{ID of associated snoRNA}
+#'   \item{Mode.of.coding}{how the SNORD is coded}
+#'   \item{SNORD.host.gene}{SNORD's host gene}
+#'   \item{Ensembl}{SNORD's Ensembl reference}
+#'   \item{Nucleotide}{the nucleotide present at the position (A,T,G or C)}
+#'   ...
+#' }
+"human_both_methylated_and_suspected"
+
 #' RiboClass from a toy dataset
 #'
 #' A riboclass containing 10 samples + 2 reference RNA.