+++ R/plot_global_profile.R
@@ -1,24 +1,31 @@
-#' Plot global profile for a given condition on a given metadata
+#' Plot global profile for given conditions on a given metadata
 #'
 #' @param ribo A RiboClass object.
 #' @param condition_col condition column for the plot
-#' @param ech modality 1 of condition_col
-#' @param ctrl modality 2 of condition_col
 #' 
 #' 
 #' @return a ggplot object
 #' @export
 #' 
 #' @examples
-#' plot_global_profile(ribo, condition_col, ech, ctrl)
+#' data("ribo_toy")
+#' data("human_methylated")
+#' ribo_toy <- remove_ribo_samples(ribo_toy,c("RNA1", "RNA2"))
+#' ribo_toy <- rename_rna(ribo_toy)  
+#' ribo_toy <- annotate_site(ribo_toy,
+#'                                 annot = human_methylated,
+#'                                 anno_rna = "rRNA",
+#'                                 anno_pos = "Position",
+#'                                 anno_value = "Nomenclature")
+#' plot_global_profile(ribo_toy, "condition")
 
-plot_global_profile <- function(ribo = ribo_adj_annot, condition_col = condition_col , ech = ech, ctrl = ctrl) {
-  
+plot_global_profile <- function(ribo = ribo_adj_annot, condition_col = NULL) { 
+
   #Check for ribo
   if (is.null(ribo)) {stop("MISSING parameter: please provide a RiboClass!")}
   if (!inherits(ribo, "RiboClass")) {stop("ribo argument is not a RiboClass!")}
   
-  metadata_x <- ribo$metadata # extract metadata from ribo object
+  meta <- ribo$metadata # extract metadata from ribo object
   ribom <- extract_data(ribo, col = "cscore", only_annotated = TRUE, position_to_rownames = TRUE)
   
   # Transform ribom into long format excluding -annotated_sites
@@ -26,51 +33,50 @@ plot_global_profile <- function(ribo = ribo_adj_annot, condition_col = condition
     tibble::rownames_to_column("annotated_sites") %>%  # Add annotated_sites as column
     tidyr::pivot_longer(cols = -annotated_sites, names_to = "samplename", values_to = "c_score")  # keep annotated_sites outside pivoting
 
-# fusion with metadata
+  # fusion with metadata
   ribom_merged <- ribom_long %>% 
-    dplyr::left_join(metadata_x, by = "samplename") # left join between metadata and ribom_long 
+    dplyr::left_join(meta, by = "samplename") # left join between metadata and ribom_long 
   
   result <- ribom_merged %>%
     dplyr::group_by(annotated_sites, !!sym(condition_col)) %>%
     dplyr::summarise(mean_c_score = mean(c_score, na.rm = TRUE),
-              sd_c_score = sd(c_score, na.rm = TRUE)) %>%
-    tidyr::pivot_wider(names_from = !!sym(condition_col), values_from = c(mean_c_score, sd_c_score))
+              sd_c_score = sd(c_score, na.rm = TRUE),
+              .groups = "drop") 
   
   #keep order as in human methylated 
-  result <- result %>% 
-    dplyr::mutate(annotated_sites = factor(annotated_sites, 
-                                    levels = unique(human_methylated$Nomenclature))) %>%
-    dplyr::arrange(annotated_sites)
+  #first order
+  fst_order <- (unique(human_methylated$Nomenclature))
   
+  # second order
+  manual <- c("5.8S_Um14","5.8S_Gm75")
+  manual <- manual[manual %in% unique(result$annotated_sites)]
+  #other_sites <- setdiff(unique(result$annotated_sites), fst_order) # sites only present in result 
   
+  #concatenation of both order
+  final_order <- c(manual, fst_order) #order 5.8S before 18 and 28_S
   
+  #check for duplicates
+  final_order <- final_order[!duplicated(final_order)]
+  #use final_order
+  result <- result %>%
+    dplyr::mutate(annotated_sites = factor(annotated_sites, levels = final_order)) 
 
-  # -----------------------------------1----------------------------------------
-  if (length(unique(condition_col)) <= 4) {   
-    # drawing the graph
-    global_profile <- ggplot(result, aes(x = annotated_sites, group = 1)) +
-      
-      geom_line(aes(y = get(paste0("mean_c_score_", ech)), color = paste0(ech))) + #PR3+
-      geom_line(aes(y = get(paste0("mean_c_score_", ctrl)), color = paste0(ctrl))) + #PR3-
-      
-      geom_ribbon(aes(ymin = get(paste0("mean_c_score_", ech)) - get(paste0("sd_c_score_", ech)), 
-                      ymax = get(paste0("mean_c_score_", ech)) + get(paste0("sd_c_score_", ech)), 
-                      fill = paste0(ech)), alpha = 0.3) +
-      
-      geom_ribbon(aes(ymin = get(paste0("mean_c_score_", ctrl)) - get(paste0("sd_c_score_", ctrl)), 
-                      ymax = get(paste0("mean_c_score_", ctrl)) + get(paste0("sd_c_score_", ctrl)),#PR3-
-                      fill = paste0(ctrl)), alpha = 0.3) +
-      
+  # draw graph
+  if (length(unique(condition_col)) > 4) stop("Too many conditions to draw a clear plot.")
+    
+    global_profile <- ggplot(result, aes(x = annotated_sites, y = mean_c_score , group = !!sym(condition_col) , color = !!sym(condition_col), fill = !!sym(condition_col))) +
+      geom_line(linewidth = 1) +
+      geom_ribbon(aes(ymin = mean_c_score - sd_c_score, ymax = mean_c_score + sd_c_score), alpha = 0.3) +
       scale_x_discrete(guide = guide_axis(angle = 90)) +
       scale_y_continuous(limits = c(0, 1)) +
-      labs(x = "RNA 2'Ome Methylation sites", y = "C-score", 
-           color = "Condition", fill = "Condition",
-           title = "2'Ome Globale Profile") +
+      labs(
+        x = "rRNA 2'Ome sites", y = "C-score",
+        color = condition_col, fill = condition_col,
+        title = "Variation in 2'Ome levels between conditions"
+      ) +
       theme_bw() +
-      scale_color_brewer(palette = "Paired")   
+      scale_color_brewer(palette = "Accent", name = "Condition") +
+      scale_fill_brewer(palette = "Accent", name = "Condition")
     
-  return(global_profile)
-  } else {
-    stop("There is to much samples into the condition to draw a clear plot")
+    return(global_profile)
   }
-}