+++ R/res_pv.R
@@ -1,4 +1,4 @@
-#' Compute statistics for c-score based on condition column
+#' Compute statistics (p-value for, Welch, Wilcoxon, Kruskal, ANOVA test) for C-score based on condition column
 #'
 #' @param test statistic test wanted student, anova or kruskal
 #' @param condition_col column condition in metadata
@@ -7,12 +7,10 @@
 #' @returns a dataframe
 #' @export
 #'
-#' @examples res_pv(ribo = ribo_adj_annot, test_name = t.test, condition_col = col)
-res_pv <- function(ribo = ribo, test = "student", condition_col = NULL) {
-  # Test arg verification
-  if(!test %in% c("student", "anova", "kruskal")) {
-    stop("Test has to be 'student', 'anova' ou 'kruskal'")
-  }
+#' @examples 
+#' data("ribo_toy")
+#' res_pv(ribo = ribo_toy, test = "student", condition_col = "condition")
+res_pv <- function(ribo = ribo, test = NULL, condition_col = NULL) {
   
   # Extract data
   data <- extract_data(ribo, only_annotated = TRUE, position_to_rownames = TRUE)
@@ -26,18 +24,94 @@ res_pv <- function(ribo = ribo, test = "student", condition_col = NULL) {
   data <- data %>%
     dplyr::left_join(ribo$metadata, by = "samplename") # left join between metadata and ribom_long
   
+  # Determine test if not provided
+  if (is.null(test)) {
+    n_conditions <- length(unique(data[[condition_col]]))
+    test <- if (n_conditions == 2) "student" 
+    else if (n_conditions > 2) "anova"
+    else message("Test can't be done with one group only")
+  }
+  
+  # Test argument verification
+  if (!test %in% c("student", "anova", "kruskal", "wilcoxon")) {
+    stop("Test must be one of: 'student', 'anova', 'kruskal' or 'wilcoxon'")
+  }
+  
+  #----------------------------------------------------------------------------
+  #             Verify white space
+  #----------------------------------------------------------------------------             
+  if (any(grepl("\\s", ribo$metadata$samplename))) {
+    stop(" Error samplename has whitespaces. Please rename your samplename.")
+  }
+  #----------------------------------------------------------------------------
+  
   # Stats
+  #create the table res_pv
+  if (test %in% c("anova", "kruskal")) {
   res_pv <- data %>%
     dplyr::group_by(annotated_sites) %>%
     dplyr::summarise(
-      p_value = ifelse(test == "student", t.test(c_score ~ get(condition_col))$p.value, # Welch test because var.equal is false (default)
-                       ifelse(test == "anova", anova(lm(c_score ~ get(condition_col)))$"Pr(>F)"[1],
-                              kruskal.test(c_score ~ get(condition_col))$p.value)),
-      fold_change = mean(c_score[get(condition_col) == unique(get(condition_col))[1]]) /
-        mean(c_score[get(condition_col) == unique(get(condition_col))[2]])
+      p_value = {
+        # Check replicats
+        rep_table <- dplyr::cur_data() # extract sub data
+        groups_table <- table(rep_table[[condition_col]])
+        # Verify number of group >= 2 replicats number >= 2
+        if (length(groups_table) >= 2 && all(groups_table >= 2)) {
+          if (test == "anova") { # test == "anova"
+            anova(lm(c_score ~ get(condition_col), data = rep_table))$"Pr(>F)"[1]
+          } else { # test == "kruskal"
+            kruskal.test(c_score ~ get(condition_col), data = rep_table)$p.value
+          }
+        } else {
+          message(" Site ", unique(rep_table$annotated_sites),
+                  " ignored : doesn't have at least 2 replicats. Help : Check for remaining 'reference' samples")
+          NA_real_
+        }
+      }
     )
+  } else { #if not anova, nor kruskal  
+    res_pv <- data %>%
+      dplyr::group_by(annotated_sites) %>%
+      #-------------------------------------------------------------------------
+      dplyr::summarise(
+        p_value = {
+          # Check replicats
+          rep_table <- dplyr::cur_data() # extract sub data
+          groups_table <- table(rep_table[[condition_col]])
+          # Verify number of group = 2 replicats number >= 2
+          if (length (groups_table) == 2 && all(groups_table >= 2)) {
+            if (test == "student") { # Welch here
+              t.test(c_score ~ get(condition_col), var.equal = FALSE)$p.value #, var.equal = FALSE for Welch test, otherwise its student test
+            } else if (test == "wilcoxon") {
+                wilcox.test(c_score ~ get(condition_col), exact = FALSE)$p.value # Mann–Whitney U test because paired = FALSE (default). exact = FALSE avoid warning when there is ex-aequo in c-score that don't enable to compute exact p-value
+              } else {
+                  NA_real_
+                }
+        } else {
+          message("Site ", unique(rep_table$annotated_sites), " ignored : doesn't have at least 2 replicats.")
+          NA_real_
+          }
+        },
+        delta_c_score = {
+          rep_table <- dplyr::cur_data()
+          groups_table <- table(rep_table[[condition_col]])
+          if (length(groups_table) == 2 && all(groups_table >= 2)) {
+            abs(mean(rep_table$c_score[rep_table[[condition_col]] == unique(rep_table[[condition_col]])[1]]) -
+                  mean(rep_table$c_score[rep_table[[condition_col]] == unique(rep_table[[condition_col]])[2]]))
+          } else {
+            NA_real_
+          }
+        }
+      )
+}
+      #-------------------------------------------------------------------------
+            
+  #           abs(mean(c_score[get(condition_col) == unique(get(condition_col))[1]]) -
+  #   mean(c_score[get(condition_col) == unique(get(condition_col))[2]])) # delta C-score for student, wilcoxon and kruskal only
+  # )
+  # }
   
-  # pval_adj
+  # pval_adj for all tests
   res_pv$p_adj <- p.adjust(res_pv$p_value, method = "fdr")
   
   return(res_pv)